+++ src/Common/JK.Common/Converters/DistanceConverter.cs
@@ -1,0 +1,51 @@
+﻿using JK.Common.Enumerations;
+using System;
+using System.Collections.Generic;
+
+namespace JK.Common.Converters
+{
+    public class DistanceConverter
+    {
+        private readonly IDictionary<Tuple<DistanceUnit, DistanceUnit>, decimal> strategies;
+
+        public DistanceConverter()
+        {
+            this.strategies = new Dictionary<Tuple<DistanceUnit, DistanceUnit>, decimal>
+            {
+                // Convert Centimeter to New Unit Of Measure
+                { Tuple.Create(DistanceUnit.Centimeters, DistanceUnit.Feet), 0.0328084m },
+                { Tuple.Create(DistanceUnit.Centimeters, DistanceUnit.Inches), 0.393701m },
+                { Tuple.Create(DistanceUnit.Centimeters, DistanceUnit.Meters), 0.01m },
+                // Convert Feet to New Unit Of Measure
+                { Tuple.Create(DistanceUnit.Feet, DistanceUnit.Centimeters), 30.48m },
+                { Tuple.Create(DistanceUnit.Feet, DistanceUnit.Inches), 12m },
+                { Tuple.Create(DistanceUnit.Feet, DistanceUnit.Meters), 0.3048m },
+                // Convert Inches to New Unit Of Measure
+                { Tuple.Create(DistanceUnit.Inches, DistanceUnit.Centimeters), 2.54m },
+                { Tuple.Create(DistanceUnit.Inches, DistanceUnit.Feet), 0.08333333m },
+                { Tuple.Create(DistanceUnit.Inches, DistanceUnit.Meters), 0.0254m },
+                // Convert Meters to New Unit of Measure
+                { Tuple.Create(DistanceUnit.Meters, DistanceUnit.Centimeters), 100m },
+                { Tuple.Create(DistanceUnit.Meters, DistanceUnit.Feet), 3.28084m },
+                { Tuple.Create(DistanceUnit.Meters, DistanceUnit.Inches), 39.3701m }
+            };
+        }
+
+        public decimal Convert(decimal originalDistance, DistanceUnit originalUnit, DistanceUnit newUnit)
+        {
+            if (originalUnit == newUnit)
+            {
+                return originalDistance;
+            }
+
+            var tuple = Tuple.Create(originalUnit, newUnit);
+            if (!this.strategies.ContainsKey(tuple))
+            {
+                throw new NotSupportedException("Unit pairing is not supported.");
+            }
+
+            var conversionFactor = this.strategies[tuple];
+            return originalDistance * conversionFactor;
+        }
+    }
+}